+++ SGL/src/SGL/System/Battery.cpp
@@ -5,6 +5,6 @@ namespace sgl {
 		int secs, pct;
 		SDL_PowerState state = SDL_GetPowerInfo(&secs, &pct);
 
-		return Battery{secs, pct, static_cast<Battery::State>(state)};
+		return Battery{secs, static_cast<int8>(pct), static_cast<Battery::State>(state)};
 	}
 }
